simple_ping_fn/requirements.txt
# Python 3.12
environs
psycopg2-binary
Flask
functions-framework
google-cloud-storage
cloud-sql-python-connector[pg8000]
